Change types int to uint16 and reorganize structs to limit padding. (the struct Neuron went from 56 bytes to 40)

+++ env_settings.h
@@ -7,7 +7,7 @@
  */
 struct env_settings
 {
-    const unsigned MaxNeuron;
+    const uint16_t MaxNeuron;
     const double fitnessGoal;
 } ENV_SETTINGS = { 200, 14.0 };
